Add the beginning of exercise 2 for the Edit component
