+++ INSTRUCTIONS.md
@@ -1,0 +1,12 @@
+# Edit Component
+
+The Edit component of a block has its interactivity.
+
+We'll add the controls for the block to the <InspectorControls>.
+
+By rendering components inside <InspectorControls>, they appear in the inspector.
+
+## Exercise
+
+### File
+- `js/src/edit.exercise.js`
