+++ js/src/save.js
@@ -10,20 +10,12 @@ import * as React from 'react';
 import { useBlockProps } from '@wordpress/block-editor';
 
 /**
- * Internal dependencies
- */
-import ProgressIndicator from './progress-indicator';
-
-/**
  * The component to save the markup.
  *
- * @param {{attributes: import('./index').Attributes}} props
  * @return {React.ReactElement} The component.
  */
-export default function Save( { attributes } ) {
+export default function Save() {
 	const blockProps = useBlockProps.save();
 
-	return <div { ...blockProps }>
-		<ProgressIndicator attributes={ attributes } />
-	</div>;
+	return <div { ...blockProps } />;
 }
